rename tabs to pages
