+++ client/src/app/auth/pages/dashboard/accounts/accounts-configuration/accounts-configuration-card.tsx
