+++ client/src/app/auth/pages/dashboard/accounts/accounts-configuration/delete-account.tsx
