+++ client/src/app/auth/pages/dashboard/accounts/accounts-configuration/deleted-account-card.tsx
