+++ client/src/app/auth/pages/dashboard/accounts/accounts-configuration/deleted-accounts-cards.tsx
