+++ client/src/app/auth/pages/dashboard/dashboard.tsx
