+++ client/src/app/auth/pages/dashboard/net-worth/net-worth-item.tsx
