+++ client/src/app/auth/pages/dashboard/spending-trends/spending-trends-chart.tsx
