+++ client/src/app/auth/pages/dashboard/sync-account-button.tsx
