+++ client/src/app/auth/pages/dashboard/welcome-header.tsx
