+++ client/src/app/auth/pages/goals/cells/editable-goal-name-cell.tsx
