+++ client/src/app/auth/pages/goals/goal-details.tsx
