+++ client/src/app/auth/pages/page-content.tsx
