+++ client/src/app/auth/pages/transactions/transactions-configuration/deleted-transaction-card.tsx
