+++ client/src/app/auth/pages/transactions/transactions-configuration/deleted-transactions-accordion.tsx
