+++ client/src/app/auth/pages/transactions/transactions-configuration/transactions-configuration.tsx
